refactor: extract shared DnD helpers for parameter reordering
Deduplicates DraggableItem and DroppableArea components into a shared
module and uses Mantine Box instead of div.
